requirements.txt
django==2.2
django-filter==2.1.0
django-graphql-jwt==0.2.1
djangorestframework==3.9.4
flake8==3.7.7
graphene-django==2.2.0
isort==4.3.17
PyYAML==5.1
